chore: resolve cli-extract-002 and track progress
